Create user_sessions schema package and refactor schema to use.
This change sets up a modular pattern where the query type and
mutation type can both inherit from subquery and submutation types
defined elsewhere.
