+++ graphene-django-backend/project/gql_platform/graphene/__init__.py
@@ -1,0 +1,5 @@
+'''
+For clarification on the boilerplate used throughout this package, first read graphene docs.
+Then read at docs.graphene-python.org.
+You could also read about models under django.
+'''
